+++ src/App/Routes.tsx
@@ -28,8 +28,8 @@ import Compartments from '../containers/Compartments';
 import ConnectedHierarchichalDataTable from '../containers/HierarchichalDataTable';
 import ConnectedLogFace from '../containers/LogFace';
 import Analysis from '../containers/pages/Analysis';
-import Home from '../containers/pages/Home/Home';
-import PregnancyHome from '../containers/pages/PregnancyHome/PregnancyHome';
+import Home from '../containers/pages/Home';
+import PregnancyHome from '../containers/pages/PregnancyHome';
 import ConnectedPatientDetails from '../containers/PatientDetails';
 import { headerShouldNotRender, oAuthUserInfoGetter } from '../helpers/utils';
 import './App.css';
